+++ src/main/java/com/ametrinstudios/ametrin/world/item/helper/ItemRegisterHelper.java
@@ -7,8 +7,7 @@ import com.ametrinstudios.ametrin.world.item.CustomBoatItem;
 
 import java.util.function.Supplier;
 
-// TODO: rename -> ItemRegisterHelper
-public class ItemRegistryHelper {
+public class ItemRegisterHelper {
     public static Supplier<CustomBoatItem> chestBoat(CustomBoatType type) {return boat(type, BoatVariants.CHEST);}
     public static Supplier<CustomBoatItem> boat(CustomBoatType type) {return boat(type, BoatVariants.DEFAULT);}
     public static Supplier<CustomBoatItem> boat(CustomBoatType type, BoatVariant<?> variant) {return ()-> new CustomBoatItem(type, variant);}